requirements.txt

pyautogen[websurfer]
pyfiglet
semantic-router
VoiceProcessingToolkit
PyMuPDF
elevenlabs
